MCP_notion-antigravity/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "antigravity-content-engine"
version = "0.1.0"
description = "Notion-native AI content engine with MCP, pipelines, and operations tooling."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "anthropic",
  "feedparser",
  "google-genai",
  "httpx",
  "mcp",
  "notion-client",
  "numpy",
  "pandas",
  "plotly",
  "python-dateutil",
  "python-dotenv",
  "pytz",
  "seaborn",
  "streamlit",
  "tweepy",
  "wordcloud",
  "yfinance",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
antigravity-mcp = "antigravity_mcp.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
